refactor: move ClientNamespace to lib/interfaces
