+++ lib/index.ts
@@ -1,6 +1,3 @@
-/**
- * v1.0.0 2021-06-17 18:34
- */
 export { RedisModule } from './redis/redis.module';
 export { DEFAULT_REDIS_CLIENT } from './redis/redis.constants';
 export { RedisService } from './redis/redis.service';
@@ -9,9 +6,9 @@ export { RedisHealthIndicator } from './redis/redis.health';
 
 // * Types & Interfaces
 export {
-    ClientNamespace,
     RedisModuleOptions,
     RedisModuleAsyncOptions,
     RedisOptionsFactory,
     RedisPingCheckOptions
 } from './redis/interfaces';
+export { ClientNamespace } from './interfaces';
